+++ src/tests/unit/tests/common/components/gear-options-button-component.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { GearMenuButton, GearOptionsButtonComponentProps } from 'common/components/gear-options-button-component';
+import { GearMenuButton, GearOptionsButtonComponentProps } from 'common/components/gear-menu-button';
 import { DropdownClickHandler } from 'common/dropdown-click-handler';
 import { FeatureFlags } from 'common/feature-flags';
 import { shallow } from 'enzyme';
